// src/ui/next/components/wizard/types.ts
export interface WizardState {
    businessType: string;
    name: string;
    desc: string;
    sellingCats: string[];
    payment: string;
    adminName: string;
    adminEmail: string;
    adminPass: string;
    template: string;
    colors: string[];
    logo: string;
    products: any[];
    domain: string;
    agents: string[];
    agentSchedule: number;
    agentTone: string;
    agentFocus: string[];
}